File/Line Length - DB lock

+++ tools/u4c/u4c.py
@@ -7,6 +7,7 @@ Design Assumptions:
 # Python Modules
 #---------------------------------------------------------------------------------------------------
 import csv
+import datetime
 import os
 import sys
 import time
@@ -19,9 +20,10 @@ import understand
 #---------------------------------------------------------------------------------------------------
 # Knowlogic Modules
 #---------------------------------------------------------------------------------------------------
-import ProjFile
+import ProjFile as PF
 import ViolationDb
 
+from tools.u4c import u4cDbWrapper as udb
 from tools.u4c import U4cFileTemplates
 from tools.ToolMgr import ToolSetup, ToolManager
 from utils.DateTime import DateTime
@@ -49,7 +51,7 @@ class U4cSetup( ToolSetup):
     def __init__( self, projFile):
         """ Handle all U4c setup
         """
-        assert( isinstance( projFile, ProjFile.ProjectFile))
+        assert( isinstance( projFile, PF.ProjectFile))
         self.projFile = projFile
         self.projRoot = projFile.paths['ProjectRoot']
 
@@ -68,8 +70,9 @@ class U4cSetup( ToolSetup):
         cmdFilePath = os.path.join( self.projToolRoot, eU4cCmdFileName)
 
         srcRoots = self.projFile.paths['SrcCodeRoot']
+        excludeDirs = self.projFile.exclude['Dirs']
         excludeFiles = self.projFile.exclude['Files_U4c']
-        srcIncludeDirs, srcCodeFiles = self.GetSrcCodeFile( srcRoots, ['.c', '.h'], excludeFiles)
+        incDirs, srcFiles = self.projFile.GetSrcCodeFiles( srcRoots, ['.c','.h'], excludeDirs, excludeFiles)
 
         # U4c Option definitions
         options = {}
@@ -84,9 +87,9 @@ class U4cSetup( ToolSetup):
                                  os.path.join( self.projToolRoot, eSrcFilesName),
                                  '\n'.join(options))
         self.CreateFile( eU4cCmdFileName, cmdContents)
-        self.CreateFile( eSrcFilesName, '\n'.join(srcCodeFiles))
+        self.CreateFile( eSrcFilesName, '\n'.join(srcFiles))
 
-        self.fileCount = len( srcCodeFiles)
+        self.fileCount = len( srcFiles)
 
     #-----------------------------------------------------------------------------------------------
     def ConvertOptions( self, options):
@@ -123,9 +126,9 @@ class U4cSetup( ToolSetup):
 #---------------------------------------------------------------------------------------------------
 class U4c( ToolManager):
     def __init__(self, projFile):
-        assert( isinstance( projFile, ProjFile.ProjectFile))
+        assert( isinstance( projFile, PF.ProjectFile))
         self.projFile = projFile
-        self.projRoot = projFile.paths['ProjectRoot']
+        self.projRoot = projFile.paths[PF.ePathProject]
 
         ToolManager.__init__(self, self.projRoot)
 
@@ -135,13 +138,14 @@ class U4c( ToolManager):
 
     #-----------------------------------------------------------------------------------------------
     def IsReadyToAnalyze(self):
-        """ we are about to create the db and put stuff in it make sure it is not locked by some
+        """ we are about to create the db and put stuff in it. Make sure it is not locked by some
             with U4c GUI open
         """
         # TODO: if U4C fixes the db open we cna use that to findout if some one has the Db open
         #       right now we will try to delete it, and hopefully get an exception
         try:
-            os.remove( self.dbName)
+            if os.path.isfile( self.dbName):
+                os.remove( self.dbName)
             status = True
         except OSError:
             status = False
@@ -162,54 +166,145 @@ class U4c( ToolManager):
     def MonitorAnalysis(self):
         """ thread that does the monitoring and completes the analysis
         """
-        fileList = []
-        analyzing = False
-        fileCount = 0
-        while self.AnalyzeActive():
-            for line in self.job.stdout:
-                line = line.decode(encoding='windows-1252').strip()
-                if line == 'Analyze':
-                    analyzing = True
-                    fileCount = len( fileList)
-                elif line.find( 'File: ') != -1:
-                    line = line.replace('File: ', '').replace(' has been added.', '')
-                    fileList.append(line)
-                else:
-                    if line in fileList:
-                        fileList.remove( line)
-                        self.analysisPercentComplete = 50.0 - ((len(fileList)/float(fileCount)*100.0)/2.0)
+        try:
+            fileList = []
+            analyzing = False
+            fileCount = 0
+            self.analysisStep = 'Analyzing Files'
+            while self.AnalyzeActive():
+                for line in self.job.stdout:
+                    line = line.decode(encoding='windows-1252').strip()
+                    if line == 'Analyze':
+                        analyzing = True
+                        fileCount = len( fileList)
+                    elif line.find( 'File: ') != -1:
+                        line = line.replace('File: ', '').replace(' has been added.', '')
+                        fileList.append(line)
+                    else:
+                        if line in fileList:
+                            fileList.remove( line)
+                            v = 100 - (len(fileList)/float(fileCount)*100.0)
+                            self.AnalysisStatusMsg( v)
 
-        # now run the review of the u4c DB which is the other half of this process
+            # now run the review of the u4c DB which is the other half of this process
+            # get the file list to check
+            srcRoots = self.projFile.paths[PF.ePathSrcRoot]
+            excludeDirs = self.projFile.exclude[PF.eExcludeDirs]
+            excludeFiles = self.projFile.exclude[PF.eExcludeU4c]
+            x, self.srcFiles = self.projFile.GetSrcCodeFiles(srcRoots, ['.h','.c'],
+                                                             excludeDirs, excludeFiles)
+
+            # open the Violation DB
+            self.vDb = ViolationDb.ViolationDb( self.projRoot)
+            self.vDb.DebugState( 1)
+
+            self.updateTime = datetime.datetime.today()
+            self.udb = udb.U4cDb( self.dbName)
+            # 12.5 % Each
+            self.analysisStep = 'Acquire DB Lock'
+            self.AnalysisStatusMsg( 0)
+            self.projFile.dbLock.acquire()
+
+            self.analysisStep = 'Metric'
+            self.AnalysisStatusMsg( 0)
+            self.MetricChecks()
+            #self.FormatChecks()
+            #self.NamingChecks()
+            #self.LanguageRestriction()
+            self.vDb.Close()
+
+            self.AnalysisStatusMsg( 100)
+            self.projFile.dbLock.release()
+        except:
+            self.vDb.Close()
+            raise
+
+    #-----------------------------------------------------------------------------------------------
+    def MetricChecks( self):
+        self.LenghtChecks()
+
+    #-----------------------------------------------------------------------------------------------
+    def LenghtChecks(self):
+        """ This function verifies all of the length limits are met on a file by fial basis.
+        """
+        fileLimit = self.projFile.metrics[PF.eMetricFile]
+        funcLimit = self.projFile.metrics[PF.eMetricFunc]
+
+        counter = 0
+        totalFiles = len(self.srcFiles)
+        for i in self.srcFiles:
+            counter += 1
+            self.AnalysisStatusMsg( (counter/float(totalFiles)*100))
+
+            # compute the relative path from a srcRoot
+            for sr in self.projFile.paths[PF.ePathSrcRoot]:
+                fn = i.replace(sr, '')
+                if fn[0] == r'\\': fn = fn[1:]
+            rpfn = fn
+            fn = os.path.split(rpfn)[1]
+
+            # Get file information
+            if os.path.splitext( fn) == '.c':
+                funcInfo = self.udb.GetFuncInfo( fn)
+            else:
+                funcInfo = {}
+            func = 'N/A'
+
+            f = open(i, 'r')
+            lines = f.readlines()
+            f.close()
+
+            # check to file size violation
+            fileSize = len(lines)
+            if fileSize > fileLimit:
+                line = fileSize
+                severity = 'Error'
+                violationId = 'Metric-File'
+                desc = 'File Length Exceeded: %s' % (fn)
+                details = 'Total line count (%d) exceeds project maximum (%d)' % (fileSize,
+                                                                                  fileLimit)
+                self.vDb.Insert(rpfn, func, severity, violationId, desc,
+                                details, line, 'U4C', self.updateTime)
+                self.vDb.Commit()
+
+            # check the lenght of each line
+            self.CheckLineLength( rpfn, funcInfo, lines)
 
 
     #-----------------------------------------------------------------------------------------------
-    def LoadDb( self):
-        """ We should now have a clean PcLint output to load into the DB
-        """
-        # open the source file
-        finName = os.path.join( self.projToolRoot, eResultFile)
+    def CheckLineLength(self, rpfn, funcInfo, lines):
+        # check all line lengths
+        fileSize = len(lines)
+        fn = os.path.split(rpfn)[1]
+        lineLimit = self.projFile.metrics[PF.eMetricLine]
+        func = 'N/A'
 
-        # open the PCLint DB
-        sl3 = ViolationDb.ViolationDb( self.projRoot)
-        sl3.DebugState( 1)
+        for lineNumber in range(0,fileSize):
+            txt = lines[lineNumber].rstrip()
+            lineLen = len(txt)
+            if lineLen > lineLimit:
+                severity = 'Error'
+                violationId = 'Metric-Line'
+                desc = 'Line Length in %s line %d' % (fn, lineNumber)
+                details = '%3d: %s' % (lineLen, txt.strip())
+                self.vDb.Insert( rpfn, func, severity, violationId, desc,
+                                 details, lineNumber, 'U4C', self.updateTime)
+        self.vDb.Commit()
 
-        # move to the DB
-        lintLoader = LintLoader( finName, sl3)
-        lintLoader.RemoveDuplicate()
-        removed, updateTime = lintLoader.InsertDb( lintLoader.reducedData)
+    #-----------------------------------------------------------------------------------------------
+    def FormatChecks(self):
+        pass
 
-        # print stats
-        print( 'Inserted %5d New' % sl3.insertNew)
-        print( 'Updated  %5d Records' % sl3.insertUpdate)
-        print( 'Select Errors: %d' % sl3.insertSelErr)
-        print( 'Insert Errors: %d' % sl3.insertInErr)
-        print( 'Update Errors: %d' % sl3.insertUpErr)
+    #-----------------------------------------------------------------------------------------------
+    def NamingChecks(self):
+        pass
 
-        # remove the old PcLint violations
-        s = "delete from violations where lastReport != ? and detectedBy = 'PcLint'"
-        sl3.Execute( s, (updateTime,))
-        sl3.Commit()
-        print( 'Removed %d Old Records - %s' % (removed, updateTime))
+    #-----------------------------------------------------------------------------------------------
+    def LanguageRestriction(self):
+        # excluded functions
+        # restricted functions
+        pass
+
 
 #========================================================================================================================
 import socket
@@ -255,8 +350,8 @@ def TestCreate():
     u4cs = U4cSetup( projRoot)
     u4cs.CreateProject( srcFiles, options)
 
-def TestRun():
 
+def TestRun():
     start = DateTime.today()
     tool = U4c( projRoot)
     tool.Analyze()
@@ -267,3 +362,6 @@ def TestRun():
     end = DateTime.today()
     delta = end - start
     print('U4C Processing: %s' % delta)
+
+if __name__ == '__main__':
+    TestRun()
